Adding the plot for the anual cycles of 'urbdyn'
Computing a scalar index from which summarize the impact of cities in atmospheric circulation.
Using monthly spatial means around the city as reference, the urbdyn index compute the spatial mean of anomalies of the 2x2 grid-box around the city center up to 850 hPa of 3-dimensional data.
It is assumed that if the city does not impact the atmospheric cirulation, around the city, the spatial anomaly of 3-dimensional variables has a zero mean